Link Travis CI build to codeclimate
We have previously added the gem from codeclimate which generates code coverage stats, and we have added the codeclimate configuration file. However to get test coverage reported we need to pass the stats to codeclimate somehow.

This changes updates the .travis.yml to tell it to forward these stats after a build to codeclimate, and authenticates by using an ecrypted API also included in the file.