Update dependencies (#18)
This project's dependencies have not been updated in a while and it's causing security warnings in projects that use it. Fixes include:

* Remove before_commit - this gem has been yanked entirely from RubyGems.
* Remove old CodeClimate test reporter - we no longer use this and it was giving deprecation warnings.
* Update Nokogiri

We also update the Travis config. This is more consistent with our standard config now, and removes reliance on gems that we no longer use.

And fix broken README badges.